edit the User,Trainee, Instructor Daos to take an Integer Id

+++ src/main/webapp/dao/InstructorHibernateDao.java
@@ -19,7 +19,7 @@ import pojo.Instructor;
  * 
  * @author root
  */
-public class InstructorHibernateDao extends GenericHibernateDAO<Instructor, Long> implements InstructorDao{
+public class InstructorHibernateDao extends GenericHibernateDAO<Instructor, Integer> implements InstructorDao{
 
     @Override
     public Instructor findByEmail(String email) {
